Transaction maliyet alanı yazıldı set edildi, TransactionOperationsPage.jsx sayfasında bug giderildi

+++ corso/src/main/java/com/infina/corso/dto/response/TransactionResponse.java
@@ -17,5 +17,5 @@ public class TransactionResponse {
     private int amount;
     private LocalDate transactionSystemDate;
     private Double rate;
-    private BigDecimal cost;
+    private Double cost;
 }
